backend/requirements.txt
fastapi
uvicorn
transformers
torch
sentencepiece
python-docx
nltk
chromadb
langdetect
